[spaces.rt.pdelab] move out of playground, refs #15
